+++ src/Systems/Interface/InputSystemInterface.h
@@ -3,6 +3,7 @@
 
 #include <cstdint>
 #include <SDL3/SDL.h>
+#include <InputState.h>
 #include <Systems/Interface/SystemInterface.h>
 #include <Vec.h>
 
@@ -16,4 +17,5 @@ public:
 	virtual void OnMouseButtonUp(uint8_t buttonIndex) = 0;
 
 	virtual Vec2 GetMousePosition() = 0;
+	virtual const InputState& GetInputState() = 0;
 };